Extract app into a separate file
